# perf(http): instantiate generic functions in `deno_http`, increase opt-level for some more hyper deps