--- a/Cargo.toml
+++ b/Cargo.toml
@@ -437,6 +437,16 @@ opt-level = 3
 opt-level = 3
 [profile.release.package.futures-util]
 opt-level = 3
+[profile.release.package.futures-io]
+opt-level = 3
+[profile.release.package.futures-channel]
+opt-level = 3
+[profile.release.package.futures-sink]
+opt-level = 3
+[profile.release.package.futures-task]
+opt-level = 3
+[profile.release.package.futures-core]
+opt-level = 3
 [profile.release.package.hyper]
 opt-level = 3
 [profile.release.package.miniz_oxide]
@@ -462,4 +472,20 @@ opt-level = 3
 [profile.release.package.zstd]
 opt-level = 3
 [profile.release.package.zstd-sys]
+opt-level = 3
+[profile.release.package.http]
+opt-level = 3
+[profile.release.package.http-body]
+opt-level = 3
+[profile.release.package.httparse]
+opt-level = 3
+[profile.release.package.mio]
+opt-level = 3
+[profile.release.package.fnv]
+opt-level = 3
+[profile.release.package.hyper-util]
+opt-level = 3
+[profile.release.package.httpdate]
+opt-level = 3
+[profile.release.package.deno_unsync]
 opt-level = 3