--- a/runtime/snapshot.rs
+++ b/runtime/snapshot.rs
@@ -61,9 +61,7 @@ pub fn create_runtime_snapshot(
       deno_cron::local::LocalCronHandler::new(),
     ),
     deno_napi::deno_napi::init_ops_and_esm::<Permissions>(),
-    deno_http::deno_http::init_ops_and_esm::<DefaultHttpPropertyExtractor>(
-      deno_http::Options::default(),
-    ),
+    deno_http::deno_http::init_ops_and_esm(deno_http::Options::default()),
     deno_io::deno_io::init_ops_and_esm(Default::default()),
     deno_fs::deno_fs::init_ops_and_esm::<Permissions>(fs.clone()),
     deno_os::deno_os::init_ops_and_esm(Default::default()),