--- a/runtime/snapshot_info.rs
+++ b/runtime/snapshot_info.rs
@@ -293,9 +293,7 @@ pub fn get_extensions_in_snapshot() -> Vec<Extension> {
     ),
     deno_cron::deno_cron::init_ops(deno_cron::local::LocalCronHandler::new()),
     deno_napi::deno_napi::init_ops::<Permissions>(),
-    deno_http::deno_http::init_ops::<DefaultHttpPropertyExtractor>(
-      deno_http::Options::default(),
-    ),
+    deno_http::deno_http::init_ops(deno_http::Options::default()),
     deno_io::deno_io::init_ops(Some(Default::default())),
     deno_fs::deno_fs::init_ops::<Permissions>(fs.clone()),
     deno_os::deno_os::init_ops(Default::default()),